msm: vidc: Use pre-computed frame delta values in WFD use case
In WFD use-case with frame skipping enabled, we wish to fool the core
into thinking that that the frames are coming at a consistent rate so
that we don't run into any issues with the core's rate control
mechanism.

Introduce a client-settable property that dictates what the timestamp
delta between any two frames might be.